Convert series ids to 64-bits.

+++ tsdb/index/tsi1/tsi1.go
@@ -341,7 +341,7 @@ func (p tagValueMergeElem) Deleted() bool {
 }
 
 type SeriesIDElem struct {
-	SeriesID uint32
+	SeriesID uint64
 	Deleted  bool
 	Expr     influxql.Expr
 }
